Move all source files to src and created object files to _objs
All source files are now located in src. Object files created during compilation will be placed in a directory called _objs instead of root. Makefile has been updated to accommodate this change.
