+++ src/blant-utils.h
@@ -1,7 +1,7 @@
 #ifndef BLANT_UTILS_H
 #define BLANT_UTILS_H
 
-#include "../blant.h"
+#include "blant.h"
 
 void ExtractPerm(char perm[_k], int i);
 TINY_GRAPH *TinyGraphInducedFromGraph(TINY_GRAPH *Gv, GRAPH *G, int *Varray);
