+++ src/compare_canon_maps.cpp
