+++ src/compute-alphas-NBE.c
