+++ src/fast-canon-map.c
