+++ src/libblant.c
