+++ src/magictable.cpp
