+++ src/slow-canon-maps.c
