+++ src/test-bin-data.c
